backend/requirements.txt
absl-py==1.4.0
aioredis==1.3.1
altair==5.1.2
amqp==5.1.1
annotated-types==0.6.0
anyio==4.0.0
app==0.0.1
appdirs==1.4.4
asgiref==3.5.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs>=21.4.0,<=22.2.0
audioread==3.0.1
autobahn==22.5.1
autograd==1.6.2
autograd-gamma==0.5.0
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binarize==0.0.1
black==22.3.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click>=8.1.3,<=8.1.4
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>==0.2.0
cmake==3.25.2
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
daphne==3.0.2
datadispatch==1.0.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.5
django-chatter==0.0.4
django-cors-headers==3.13.0
django-email-users==1.1.0
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-rest-swagger==0.3.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.24.0
drf-yasg==1.20.0
exceptiongroup==1.1.3
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
face-utils==0.1.0
fastapi==0.109.2
ffmpeg==1.4
filelock==3.13.1
flatbuffers==23.1.21
fonttools==4.42.1
formulaic==0.6.6
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hiredis==2.0.0
htmltools==0.3.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==6.1.0
imutils==0.5.4
incremental==21.3.0
inflection==0.5.1
interface-meta==1.3.0
ipaddress==1.0.23
ipython==8.17.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonfield==3.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kaleido==0.2.1
keras==2.11.0
kiwisolver==1.4.5
kombu==5.2.4
lazy_loader==0.3
libclang==15.0.6.1
librosa==0.10.1
lifelines==0.27.8
linkify-it-py==2.0.2
llvmlite==0.41.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.4
mypy-extensions==0.4.3
netbox==0.0.2
networkx==3.2.1
noisereduce==3.0.0
numba==0.58.1
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==2.1.1
parso==0.8.3
pathspec==0.9.0
patroni==3.2.0
Pillow==9.4.0
ping3==4.0.3
platformdirs==2.5.2
plotly==5.18.0
pooch==1.8.0
prettytable==3.9.0
prompt-toolkit>=3.0.30,<=3.0.37
protobuf>=3.20,<4
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
python-netbox==0.0.24
pytz==2022.1
PyYAML==6.0.1
redis==4.3.3
referencing==0.31.0
requests==2.28.0
requests-oauthlib==1.3.1
rich==13.6.0
routing==0.2.0
rpds-py==0.12.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.3.2
scipy==1.11.2
service-identity==21.1.0
shiny==0.5.1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.2
stack-data==0.6.3
starlette==0.36.3
streamlit==1.28.2
swapper==1.3.0
sympy==1.12
tasks==2.8.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.13.0
Twisted==22.4.0
twisted-iocpsupport==1.0.4
txaio==22.2.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
ui==0.1.4
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.23.2
validators==0.22.0
vine==5.0.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.14.1
ydiff==1.2
zipp==3.15.0
zope.interface==5.4.0
